feat: 이벤트 시스템 모듈화 및 파일 분리
- 이벤트 시스템, 핸들러, 레지스트리를 별도 파일로 분리
- UI 렌더러를 독립 모듈로 추출
