+++ src/basic/main.basic.js
@@ -24,96 +24,13 @@ import {
 import cartStore from './features/cart/cartStore.js';
 // 🛠️ 순수 유틸리티 함수들 import
 import { CartUtils } from './features/cart/cartUtils.js';
+import { registerEventListeners } from './features/events/eventRegistry.js';
 import { setupObservers } from './features/observerFactory.js';
 import createProductStore from './features/product/productStore.js';
 import { ProductUtils } from './features/product/productUtils.js';
+import uiRenderer from './features/ui/uiRenderer.js';
 import uiStore from './features/ui/uiStore.js';
-import { formatNumber, formatPrice, safeParseInt, when, whenValue } from './utils/dataUtils.js';
-import { getElement, querySelector, setInnerHTML, setStyle, setTextContent } from './utils/domUtils.js';
-
-// 🏪 UI 렌더링 모듈 (React 스타일)
-const uiRenderer = {
-  // 상태 기반 UI 업데이트
-  renderCartDisplay: (totalItems, finalTotal) => {
-    setTextContent('item-count', `🛍️ ${totalItems} items in cart`);
-
-    const totalDiv = querySelector(getElement('cart-total'), '.text-2xl');
-    if (totalDiv) {
-      totalDiv.textContent = formatPrice(finalTotal);
-    }
-  },
-
-  renderPointsDisplay: (totalPoints) => {
-    const pointsDisplay = whenValue(totalPoints > 0, `적립 포인트: ${totalPoints}p`, '적립 포인트: 0p');
-    setTextContent('loyalty-points', pointsDisplay);
-    setStyle('loyalty-points', 'display', 'block');
-  },
-
-  renderTuesdaySpecial: (isTuesday, finalTotal) => {
-    const tuesdaySpecial = getElement('tuesday-special');
-    if (tuesdaySpecial) {
-      if (isTuesday && finalTotal > 0) {
-        tuesdaySpecial.classList.remove('hidden');
-      } else {
-        tuesdaySpecial.classList.add('hidden');
-      }
-    }
-  },
-
-  renderStockMessages: (stockMessages) => {
-    const stockMsg = stockMessages.join('\n');
-    setTextContent('stock-status', stockMsg);
-  },
-
-  renderSummaryDetails: (summaryItems) => {
-    setInnerHTML('summary-details', summaryItems.join(''));
-  },
-
-  renderDiscountInfo: (totalDiscountRate, savedAmount) => {
-    const discountInfoDiv = getElement('discount-info');
-    if (totalDiscountRate > 0 && savedAmount > 0) {
-      discountInfoDiv.innerHTML = /*html*/ `
-        <div class="bg-green-500/20 rounded-lg p-3">
-          <div class="flex justify-between items-center mb-1">
-            <span class="text-xs uppercase tracking-wide text-green-400">총 할인율</span>
-            <span class="text-sm font-medium text-green-400">${(totalDiscountRate * 100).toFixed(1)}%</span>
-          </div>
-          <div class="text-2xs text-gray-300">₩${formatNumber(savedAmount)} 할인되었습니다</div>
-        </div>
-      `;
-    } else {
-      discountInfoDiv.innerHTML = '';
-    }
-  },
-
-  renderCartItemStyles: (cartItems) => {
-    Array.from(cartItems).forEach((cartItem) => {
-      const quantity = CartUtils.getQuantityFromCartItem(cartItem);
-      const priceElems = cartItem.querySelectorAll('.text-lg, .text-xs');
-
-      priceElems.forEach((elem) => {
-        if (elem.classList.contains('text-lg')) {
-          elem.style.fontWeight = whenValue(quantity >= UI_CONSTANTS.QUANTITY_THRESHOLD_FOR_BOLD, 'bold', 'normal');
-        }
-      });
-    });
-  },
-
-  renderManualOverlay: (isVisible) => {
-    const manualOverlay = getElement('manual-overlay');
-    const manualColumn = getElement('manual-column');
-
-    when(isVisible, () => {
-      manualOverlay.classList.remove('hidden');
-      manualColumn.classList.remove('translate-x-full');
-    });
-
-    when(!isVisible, () => {
-      manualOverlay.classList.add('hidden');
-      manualColumn.classList.add('translate-x-full');
-    });
-  },
-};
+import { getElement } from './utils/domUtils.js';
 
 // 🏪 데이터 변환 함수들
 const transformServerDataToClientState = (serverData) => {
@@ -302,307 +219,6 @@ const pointCalculator = {
   },
 };
 
-// 🏪 이벤트 시스템 - WeakMap, Set, Map 활용
-const eventSystem = {
-  // WeakMap: 이벤트 엘리먼트 관리 (메모리 누수 방지)
-  eventElements: new WeakMap(),
-
-  // Set: 이벤트 타입 관리 (중복 방지)
-  eventTypes: new Set(['click', 'change', 'input', 'submit', 'keydown', 'keyup', 'focus', 'blur']),
-
-  // 이벤트 타입 동적 등록
-  registerEventType: (eventType) => {
-    eventSystem.eventTypes.add(eventType);
-  },
-
-  // 이벤트 타입 동적 제거
-  unregisterEventType: (eventType) => {
-    eventSystem.eventTypes.delete(eventType);
-  },
-
-  // 이벤트 타입 존재 여부 확인
-  hasEventType: (eventType) => {
-    return eventSystem.eventTypes.has(eventType);
-  },
-
-  // 등록된 모든 이벤트 타입 조회
-  getRegisteredEventTypes: () => {
-    return Array.from(eventSystem.eventTypes);
-  },
-
-  // Map: 이벤트 핸들러 관리 (타입별 핸들러 그룹화)
-  eventHandlers: new Map(),
-
-  // 이벤트 위임을 위한 부모 요소 탐색
-  findEventTarget: (event, selector) => {
-    let target = event.target;
-    while (target && target !== event.currentTarget) {
-      if (target.matches(selector)) {
-        return target;
-      }
-      target = target.parentElement;
-    }
-    // currentTarget도 확인 (매뉴얼 오버레이 같은 경우)
-    if (event.currentTarget.matches(selector)) {
-      return event.currentTarget;
-    }
-    return null;
-  },
-
-  // 이벤트 핸들러 등록
-  registerHandler: (eventType, selector, handler) => {
-    // 이벤트 타입이 등록되지 않은 경우 자동 등록
-    if (!eventSystem.hasEventType(eventType)) {
-      eventSystem.registerEventType(eventType);
-    }
-
-    if (!eventSystem.eventHandlers.has(eventType)) {
-      eventSystem.eventHandlers.set(eventType, new Map());
-    }
-    eventSystem.eventHandlers.get(eventType).set(selector, handler);
-  },
-
-  // 이벤트 위임 핸들러 생성
-  createDelegatedHandler: (eventType) => {
-    return (event) => {
-      const handlers = eventSystem.eventHandlers.get(eventType);
-      if (!handlers) return;
-
-      for (const [selector, handler] of handlers) {
-        const target = eventSystem.findEventTarget(event, selector);
-        if (target) {
-          handler(event, target);
-          break;
-        }
-      }
-    };
-  },
-
-  // 이벤트 리스너 등록
-  attachEventListeners: (container) => {
-    eventSystem.eventElements.set(container, new Set());
-
-    for (const eventType of eventSystem.eventTypes) {
-      const delegatedHandler = eventSystem.createDelegatedHandler(eventType);
-      container.addEventListener(eventType, delegatedHandler);
-      eventSystem.eventElements.get(container).add(eventType);
-    }
-  },
-
-  // 특정 이벤트 타입 리스너 동적 등록
-  attachEventListener: (container, eventType) => {
-    if (!eventSystem.hasEventType(eventType)) {
-      eventSystem.registerEventType(eventType);
-    }
-
-    const attachedEvents = eventSystem.eventElements.get(container);
-    if (attachedEvents && !attachedEvents.has(eventType)) {
-      const delegatedHandler = eventSystem.createDelegatedHandler(eventType);
-      container.addEventListener(eventType, delegatedHandler);
-      attachedEvents.add(eventType);
-    }
-  },
-
-  // 특정 이벤트 타입 리스너 동적 제거
-  detachEventListener: (container, eventType) => {
-    const attachedEvents = eventSystem.eventElements.get(container);
-    if (attachedEvents && attachedEvents.has(eventType)) {
-      const delegatedHandler = eventSystem.createDelegatedHandler(eventType);
-      container.removeEventListener(eventType, delegatedHandler);
-      attachedEvents.delete(eventType);
-    }
-  },
-
-  // 이벤트 리스너 제거
-  detachEventListeners: (container) => {
-    const attachedEvents = eventSystem.eventElements.get(container);
-    if (attachedEvents) {
-      for (const eventType of attachedEvents) {
-        const delegatedHandler = eventSystem.createDelegatedHandler(eventType);
-        container.removeEventListener(eventType, delegatedHandler);
-      }
-      eventSystem.eventElements.delete(container);
-    }
-  },
-
-  // 모든 이벤트 핸들러 제거
-  clearHandlers: () => {
-    eventSystem.eventHandlers.clear();
-  },
-
-  // 특정 이벤트 타입의 모든 핸들러 제거
-  clearHandlersForEventType: (eventType) => {
-    eventSystem.eventHandlers.delete(eventType);
-  },
-
-  // 특정 선택자의 핸들러 제거
-  removeHandler: (eventType, selector) => {
-    const handlers = eventSystem.eventHandlers.get(eventType);
-    if (handlers) {
-      handlers.delete(selector);
-      // 핸들러가 없으면 이벤트 타입도 제거
-      if (handlers.size === 0) {
-        eventSystem.eventHandlers.delete(eventType);
-      }
-    }
-  },
-
-  // 이벤트 시스템 상태 조회
-  getEventSystemStatus: () => {
-    return {
-      registeredEventTypes: eventSystem.getRegisteredEventTypes(),
-      totalHandlers: Array.from(eventSystem.eventHandlers.entries()).map(([eventType, handlers]) => ({
-        eventType,
-        handlerCount: handlers.size,
-      })),
-    };
-  },
-};
-
-// 🏪 이벤트 핸들러 모듈
-const eventHandlers = {
-  // 매뉴얼 토글 이벤트 핸들러
-  handleManualToggle: () => {
-    uiStore.dispatch({ type: 'TOGGLE_MANUAL_OVERLAY' });
-    const isVisible = uiStore.getState().isManualOverlayVisible;
-    uiRenderer.renderManualOverlay(isVisible);
-  },
-
-  // 매뉴얼 오버레이 배경 클릭 이벤트 핸들러
-  handleManualOverlayClick: (event) => {
-    // 배경 클릭 시에만 모달 닫기 (event.target이 매뉴얼 오버레이 자체인 경우)
-    if (event.target.id === 'manual-overlay') {
-      uiStore.dispatch({ type: 'SET_MANUAL_OVERLAY_VISIBLE', payload: false });
-      uiRenderer.renderManualOverlay(false);
-    }
-  },
-
-  // 매뉴얼 닫기 버튼 이벤트 핸들러
-  handleManualClose: () => {
-    uiStore.dispatch({ type: 'SET_MANUAL_OVERLAY_VISIBLE', payload: false });
-    uiRenderer.renderManualOverlay(false);
-  },
-
-  // 장바구니 추가 이벤트 핸들러
-  handleAddToCart: () => {
-    const sel = getElement('product-select');
-    const selItem = sel.value;
-    const hasItem = productStore.getState().products.some((product) => product.id === selItem);
-
-    if (!selItem || !hasItem) {
-      return;
-    }
-
-    const itemToAdd = ProductUtils.findProductById(selItem, productStore.getState().products);
-    if (itemToAdd && itemToAdd.q > 0) {
-      const cartContainer = getElement('cart-items');
-      const existingItem = getElement(itemToAdd.id);
-
-      if (existingItem) {
-        // 기존 아이템 수량 증가
-        const currentQty = CartUtils.getQuantityFromCartItem(existingItem);
-        const newQty = currentQty + 1;
-        if (newQty <= itemToAdd.q + currentQty) {
-          CartUtils.setQuantityToCartItem(existingItem, newQty);
-          productStore.dispatch({
-            type: 'DECREASE_STOCK',
-            payload: { productId: itemToAdd.id, quantity: 1 },
-          });
-        } else {
-          alert('재고가 부족합니다.');
-          return;
-        }
-      } else {
-        // 새 아이템 추가
-        cartContainer.insertAdjacentHTML('beforeend', CartUtils.createCartItemHTML(itemToAdd));
-        productStore.dispatch({
-          type: 'DECREASE_STOCK',
-          payload: { productId: itemToAdd.id, quantity: 1 },
-        });
-      }
-
-      handleCalculateCartStuff();
-      cartStore.dispatch({ type: 'SET_LAST_SELECTED', payload: itemToAdd.id });
-    }
-  },
-
-  // 장바구니 아이템 클릭 이벤트 핸들러
-  handleCartItemClick: (event, target) => {
-    const prodId = target.dataset.productId;
-    const itemElem = getElement(prodId);
-    const prod = ProductUtils.findProductById(prodId, productStore.getState().products);
-
-    if (target.classList.contains('quantity-change')) {
-      // 수량 변경
-      const qtyChange = safeParseInt(target.dataset.change);
-      const currentQty = CartUtils.getQuantityFromCartItem(itemElem);
-      const newQty = currentQty + qtyChange;
-
-      if (newQty > 0 && newQty <= prod.q + currentQty) {
-        CartUtils.setQuantityToCartItem(itemElem, newQty);
-        productStore.dispatch({
-          type: 'DECREASE_STOCK',
-          payload: { productId: prodId, quantity: qtyChange },
-        });
-      } else if (newQty <= 0) {
-        productStore.dispatch({
-          type: 'INCREASE_STOCK',
-          payload: { productId: prodId, quantity: currentQty },
-        });
-        itemElem.remove();
-      } else {
-        alert('재고가 부족합니다.');
-      }
-    } else if (target.classList.contains('remove-item')) {
-      // 아이템 제거
-      const remQty = CartUtils.getQuantityFromCartItem(itemElem);
-      productStore.dispatch({
-        type: 'INCREASE_STOCK',
-        payload: { productId: prodId, quantity: remQty },
-      });
-      itemElem.remove();
-    }
-
-    handleCalculateCartStuff();
-    onUpdateSelectOptions();
-  },
-
-  // 장바구니 아이템 호버 이벤트 핸들러 (동적 이벤트 타입 예제)
-  handleCartItemHover: (event, target) => {
-    target.style.transform = 'scale(1.02)';
-    target.style.transition = 'transform 0.2s ease';
-  },
-
-  // 장바구니 아이템 호버 아웃 이벤트 핸들러 (동적 이벤트 타입 예제)
-  handleCartItemLeave: (event, target) => {
-    target.style.transform = 'scale(1)';
-  },
-
-  // 이벤트 리스너 등록
-  registerEventListeners: () => {
-    const appContainer = getElement('app');
-
-    // 이벤트 위임 핸들러 등록
-    eventSystem.registerHandler('click', '#manual-toggle', eventHandlers.handleManualToggle);
-    eventSystem.registerHandler('click', '#manual-overlay', eventHandlers.handleManualOverlayClick);
-    eventSystem.registerHandler('click', '#manual-close', eventHandlers.handleManualClose);
-    eventSystem.registerHandler('click', '#add-to-cart', eventHandlers.handleAddToCart);
-    eventSystem.registerHandler('click', '.quantity-change', eventHandlers.handleCartItemClick);
-    eventSystem.registerHandler('click', '.remove-item', eventHandlers.handleCartItemClick);
-
-    // 동적 이벤트 타입 등록 예제
-    eventSystem.registerEventType('mouseenter');
-    eventSystem.registerEventType('mouseleave');
-
-    // 동적으로 등록된 이벤트 타입에 대한 핸들러 등록
-    eventSystem.registerHandler('mouseenter', '.cart-item', eventHandlers.handleCartItemHover);
-    eventSystem.registerHandler('mouseleave', '.cart-item', eventHandlers.handleCartItemLeave);
-
-    // 이벤트 위임 리스너 등록
-    eventSystem.attachEventListeners(appContainer);
-  },
-};
-
 // 🧩 컴포넌트 조합 함수
 const createMainContent = () => /*html*/ `
   <div class="bg-white border border-gray-200 p-8 overflow-y-auto">
@@ -634,7 +250,7 @@ function main() {
   setupObservers(cartStore, productStore, uiStore, uiRenderer);
 
   // 이벤트 위임 리스너 등록
-  eventHandlers.registerEventListeners();
+  registerEventListeners(handleCalculateCartStuff, onUpdateSelectOptions, cartStore, productStore, uiStore);
 
   onUpdateSelectOptions();
   handleCalculateCartStuff();
@@ -777,6 +393,7 @@ function onUpdateSelectOptions() {
   // 재고 상태에 따른 스타일 적용
   sel.style.borderColor = totalStock < UI_CONSTANTS.TOTAL_STOCK_WARNING_THRESHOLD ? 'orange' : '';
 }
+
 // 📦 재고 상태 헬퍼 함수 (도메인 함수 사용)
 
 // 📊 계산 로직 함수들 - 순수 함수로 분리
